Move tree-independent geometrical functions to morphmath.
